+++ src/main/kotlin/cloud/glitchdev/rfu/gui/hud/elements/FrogcoinBlessingDisplay.kt
@@ -1,4 +1,4 @@
-﻿package cloud.glitchdev.rfu.gui.hud.elements
+package cloud.glitchdev.rfu.gui.hud.elements
 
 import cloud.glitchdev.rfu.config.categories.LotusAtollSettings
 import cloud.glitchdev.rfu.constants.fishing.FishingIslands
@@ -14,9 +14,10 @@ import kotlin.time.Duration.Companion.milliseconds
 
 @HudElement
 object FrogcoinBlessingDisplay : AbstractTextHudElement("frogcoinBlessing") {
-
-    override val enabled: Boolean
-        get() = LotusAtollSettings.frogcoinBlessingDisplay && (isEditing || (World.island == FishingIslands.ATOLL && FrogcoinBlessingFeature.activeBlessings.isNotEmpty()))
+    override val requirement: Boolean
+        get() = LotusAtollSettings.frogcoinBlessingDisplay
+    override val isElementActive: Boolean
+        get() = World.island == FishingIslands.ATOLL && FrogcoinBlessingFeature.activeBlessings.isNotEmpty()
 
     override fun onInitialize() {
         super.onInitialize()